+++ OpenSim/Region/ScriptEngine/Shared/Api/Interface/ILSL_Api.cs
@@ -230,7 +230,7 @@ namespace OpenSim.Region.ScriptEngine.Shared.Api.Interfaces
               void llLoopSoundSlave(string sound, double volume);
               void llMakeExplosion(int particles, double scale, double vel, double lifetime, double arc, string texture, LSL_Vector offset);
               void llMakeFire(int particles, double scale, double vel, double lifetime, double arc, string texture, LSL_Vector offset);
-              void llMakeFountain(int particles, float scale, float vel, float lifetime, float arc, int bounce, string texture, LSL_Vector offset, float bounce_offset);
+              void llMakeFountain(int particles, double scale, double vel, double lifetime, double arc, int bounce, string texture, LSL_Vector offset, double bounce_offset);
               void llMakeSmoke(int particles, double scale, double vel, double lifetime, double arc, string texture, LSL_Vector offset);
               void llMapDestination(string simname, LSL_Vector pos, LSL_Vector look_at);
               LSL_String llMD5String(string src, int nonce);
